# frontend/src/api/generated/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "D&D Kids Resources API",
    "description": "API for D&D 5e resource management",
    "version": "2.0.0"
  },
  "paths": {
...
    "/api/npcs": {
      "get": {
        "tags": [
          "npcs"
        ],
        "summary": "List Npcs",
        "description": "List all NPCs.",
        "operationId": "listNpcs",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 500,
              "minimum": 1,
              "default": 100,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/NPC"
                  },
                  "title": "Response Listnpcs"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "npcs"
        ],
        "summary": "Create Npc",
        "description": "Create a new NPC.",
        "operationId": "createNpc",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NPCCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NPC"
                }
              }
            }
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NpcError"
                }
              }
            },
            "description": "Bad Request"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NpcError"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/npcs/{npc_id}": {
      "get": {
        "tags": [
          "npcs"
        ],
        "summary": "Get Npc",
        "description": "Get a specific NPC by ID.",
        "operationId": "getNpc",
        "parameters": [
          {
            "name": "npc_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Npc Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NPC"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NpcError"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "npcs"
        ],
        "summary": "Update Npc",
        "description": "Update an existing NPC.",
        "operationId": "updateNpc",
        "parameters": [
          {
            "name": "npc_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Npc Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NPCUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NPC"
                }
              }
            }
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NpcError"
                }
              }
            },
            "description": "Bad Request"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NpcError"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "npcs"
        ],
        "summary": "Delete Npc",
        "description": "Delete an NPC.",
        "operationId": "deleteNpc",
        "parameters": [
          {
            "name": "npc_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Npc Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NpcError"
                }
              }
            },
            "description": "Bad Request"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NpcError"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ArmorClass": {
        "properties": {
          "value": {
            "type": "integer",
            "title": "Value"
          },
          "note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          },
          "alternatives": {
            "items": {
              "$ref": "#/components/schemas/ArmorClassEntry"
            },
            "type": "array",
            "title": "Alternatives"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "value"
        ],
        "title": "ArmorClass"
      },
...
      "CreatureType": {
        "properties": {
          "category": {
            "type": "string",
            "minLength": 1,
            "title": "Category"
          },
          "tags": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Tags"
          },
          "swarm_size": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "tiny",
                  "small",
                  "medium",
                  "large",
                  "huge",
                  "gargantuan"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Swarm Size"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "category"
        ],
        "title": "CreatureType"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "HitPoints": {
        "properties": {
          "average": {
            "type": "integer",
            "title": "Average"
          },
          "formula": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Formula"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "average"
        ],
        "title": "HitPoints"
      },
...
      "MovementSpeed": {
        "properties": {
          "mode": {
            "type": "string",
            "enum": [
              "walk",
              "burrow",
              "climb",
              "fly",
              "swim"
            ],
            "title": "Mode"
          },
          "feet": {
            "type": "integer",
            "title": "Feet"
          },
          "note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          },
          "hover": {
            "type": "boolean",
            "title": "Hover",
            "default": false
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "mode",
          "feet"
        ],
        "title": "MovementSpeed"
      },
      "NPC": {
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1,
            "title": "Name"
          },
          "race": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Race"
          },
          "gender": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Gender"
          },
          "background": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Background"
          },
          "sizes": {
            "items": {
              "type": "string",
              "enum": [
                "tiny",
                "small",
                "medium",
                "large",
                "huge",
                "gargantuan"
              ]
            },
            "type": "array",
            "title": "Sizes"
          },
          "alignment": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Alignment"
          },
          "creature_type": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/CreatureType"
              },
              {
                "type": "null"
              }
            ]
          },
          "ac": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ArmorClass"
              },
              {
                "type": "null"
              }
            ]
          },
          "hp": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/HitPoints"
              },
              {
                "type": "null"
              }
            ]
          },
          "speed": {
            "items": {
              "$ref": "#/components/schemas/MovementSpeed"
            },
            "type": "array",
            "title": "Speed"
          },
          "abilities": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AbilityScores"
              },
              {
                "type": "null"
              }
            ]
          },
          "saving_throws": {
            "additionalProperties": {
              "type": "integer"
            },
            "propertyNames": {
              "enum": [
                "str",
                "dex",
                "con",
                "int",
                "wis",
                "cha"
              ]
            },
            "type": "object",
            "title": "Saving Throws"
          },
          "skills": {
            "additionalProperties": {
              "type": "integer"
            },
            "type": "object",
            "title": "Skills"
          },
          "passive_perception": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Passive Perception"
          },
          "damage_resistances": {
            "items": {
              "$ref": "#/components/schemas/DamageModifier"
            },
            "type": "array",
            "title": "Damage Resistances"
          },
          "damage_immunities": {
            "items": {
              "$ref": "#/components/schemas/DamageModifier"
            },
            "type": "array",
            "title": "Damage Immunities"
          },
          "damage_vulnerabilities": {
            "items": {
              "$ref": "#/components/schemas/DamageModifier"
            },
            "type": "array",
            "title": "Damage Vulnerabilities"
          },
          "condition_immunities": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Condition Immunities"
          },
          "senses": {
            "items": {
              "$ref": "#/components/schemas/Sense"
            },
            "type": "array",
            "title": "Senses"
          },
          "languages": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Languages"
          },
          "features": {
            "$ref": "#/components/schemas/MonsterFeatures"
          },
          "cr": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cr"
          },
          "cr_note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cr Note"
          },
          "experience_points": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Experience Points"
          },
          "appearance": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Appearance"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "id": {
            "type": "integer",
            "title": "Id"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "name",
          "id"
        ],
        "title": "NPC"
      },
      "NPCCreate": {
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1,
            "title": "Name"
          },
          "race": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Race"
          },
          "gender": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Gender"
          },
          "background": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Background"
          },
          "sizes": {
            "items": {
              "type": "string",
              "enum": [
                "tiny",
                "small",
                "medium",
                "large",
                "huge",
                "gargantuan"
              ]
            },
            "type": "array",
            "title": "Sizes"
          },
          "alignment": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Alignment"
          },
          "creature_type": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/CreatureType"
              },
              {
                "type": "null"
              }
            ]
          },
          "ac": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ArmorClass"
              },
              {
                "type": "null"
              }
            ]
          },
          "hp": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/HitPoints"
              },
              {
                "type": "null"
              }
            ]
          },
          "speed": {
            "items": {
              "$ref": "#/components/schemas/MovementSpeed"
            },
            "type": "array",
            "title": "Speed"
          },
          "abilities": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AbilityScores"
              },
              {
                "type": "null"
              }
            ]
          },
          "saving_throws": {
            "additionalProperties": {
              "type": "integer"
            },
            "propertyNames": {
              "enum": [
                "str",
                "dex",
                "con",
                "int",
                "wis",
                "cha"
              ]
            },
            "type": "object",
            "title": "Saving Throws"
          },
          "skills": {
            "additionalProperties": {
              "type": "integer"
            },
            "type": "object",
            "title": "Skills"
          },
          "passive_perception": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Passive Perception"
          },
          "damage_resistances": {
            "items": {
              "$ref": "#/components/schemas/DamageModifier"
            },
            "type": "array",
            "title": "Damage Resistances"
          },
          "damage_immunities": {
            "items": {
              "$ref": "#/components/schemas/DamageModifier"
            },
            "type": "array",
            "title": "Damage Immunities"
          },
          "damage_vulnerabilities": {
            "items": {
              "$ref": "#/components/schemas/DamageModifier"
            },
            "type": "array",
            "title": "Damage Vulnerabilities"
          },
          "condition_immunities": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Condition Immunities"
          },
          "senses": {
            "items": {
              "$ref": "#/components/schemas/Sense"
            },
            "type": "array",
            "title": "Senses"
          },
          "languages": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Languages"
          },
          "features": {
            "$ref": "#/components/schemas/MonsterFeatures"
          },
          "cr": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cr"
          },
          "cr_note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cr Note"
          },
          "experience_points": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Experience Points"
          },
          "appearance": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Appearance"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "name"
        ],
        "title": "NPCCreate"
      },
      "NPCUpdate": {
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1,
            "title": "Name"
          },
          "race": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Race"
          },
          "gender": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Gender"
          },
          "background": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Background"
          },
          "sizes": {
            "items": {
              "type": "string",
              "enum": [
                "tiny",
                "small",
                "medium",
                "large",
                "huge",
                "gargantuan"
              ]
            },
            "type": "array",
            "title": "Sizes"
          },
          "alignment": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Alignment"
          },
          "creature_type": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/CreatureType"
              },
              {
                "type": "null"
              }
            ]
          },
          "ac": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ArmorClass"
              },
              {
                "type": "null"
              }
            ]
          },
          "hp": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/HitPoints"
              },
              {
                "type": "null"
              }
            ]
          },
          "speed": {
            "items": {
              "$ref": "#/components/schemas/MovementSpeed"
            },
            "type": "array",
            "title": "Speed"
          },
          "abilities": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AbilityScores"
              },
              {
                "type": "null"
              }
            ]
          },
          "saving_throws": {
            "additionalProperties": {
              "type": "integer"
            },
            "propertyNames": {
              "enum": [
                "str",
                "dex",
                "con",
                "int",
                "wis",
                "cha"
              ]
            },
            "type": "object",
            "title": "Saving Throws"
          },
          "skills": {
            "additionalProperties": {
              "type": "integer"
            },
            "type": "object",
            "title": "Skills"
          },
          "passive_perception": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Passive Perception"
          },
          "damage_resistances": {
            "items": {
              "$ref": "#/components/schemas/DamageModifier"
            },
            "type": "array",
            "title": "Damage Resistances"
          },
          "damage_immunities": {
            "items": {
              "$ref": "#/components/schemas/DamageModifier"
            },
            "type": "array",
            "title": "Damage Immunities"
          },
          "damage_vulnerabilities": {
            "items": {
              "$ref": "#/components/schemas/DamageModifier"
            },
            "type": "array",
            "title": "Damage Vulnerabilities"
          },
          "condition_immunities": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Condition Immunities"
          },
          "senses": {
            "items": {
              "$ref": "#/components/schemas/Sense"
            },
            "type": "array",
            "title": "Senses"
          },
          "languages": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Languages"
          },
          "features": {
            "$ref": "#/components/schemas/MonsterFeatures"
          },
          "cr": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cr"
          },
          "cr_note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cr Note"
          },
          "experience_points": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Experience Points"
          },
          "appearance": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Appearance"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "name"
        ],
        "title": "NPCUpdate"
      },
      "NpcError": {
        "properties": {
          "code": {
            "type": "string",
            "enum": [
              "npc_not_found",
              "failed_to_create_npc",
              "failed_to_update_npc",
              "failed_to_delete_npc"
            ],
            "title": "Code"
          },
          "message": {
            "type": "string",
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "title": "NpcError"
      },
...
    }
  }
}